--- public/data/mock.js
@@ -1,4 +1,0 @@
-const users = [
-  { account: 'peanutbutterfesta', name: 'namju' },
-  { account: 'bitlover', name: 'kiman' },
-];
