+++ src/pages/jeongdo/Main/CommentItem/CommentItem.js
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import './Comment.scss';
 
-const CommentItem = ({ comment, comments, id, idx, setComments }) => {
+const CommentItem = ({ comment, comments, setComments, id }) => {
   const [isLike, setIsLike] = useState(false);
   const likeClick = e => {
     setIsLike(!isLike);
@@ -11,15 +11,15 @@ const CommentItem = ({ comment, comments, id, idx, setComments }) => {
     setComments(comments.filter(commentGuest => commentGuest.id !== id));
   };
   return (
-    <div className="post__comment__guest" key={idx}>
+    <div className="post__comment__guest">
       <div className="post__comment__guest-info">
-        <span>{comment.name}</span>
-        <p>{comment.comment}</p>
+        <span>{comment.userName}</span>
+        <p>{comment.content}</p>
       </div>
       <div className="post__comment__guest-icon">
-        <i class="far fa-trash-alt" onClick={() => handleDelete(id)} />
+        <i className="far fa-trash-alt" onClick={() => handleDelete(id)} />
         {isLike ? (
-          <i class="fas fa-heart on" onClick={likeClick} />
+          <i className="fas fa-heart on" onClick={likeClick} />
         ) : (
           <i className="far fa-heart heart " onClick={likeClick} />
         )}
